move admin comps to /Admin as it will use Context directly
