+++ src/pages/Admin/Charity/EditProfile/index.tsx
@@ -3,10 +3,10 @@ import { FormProvider, useForm } from "react-hook-form";
 import { FlatFormValues, FormValues } from "./types";
 import { Profile, endow } from "services/types";
 import { useProfileQuery } from "services/aws/aws";
-import { FormError, FormSkeleton } from "components/admin";
 import { adminRoutes } from "constants/routes";
 import { unsdgs } from "constants/unsdgs";
 import { useAdminContext } from "../../Context";
+import { FormError, FormSkeleton } from "../../components";
 import Seo from "../Seo";
 import Form from "./Form";
 import { getEndowDesignationLabelValuePair } from "./getEndowDesignationLabelValuePair";
