+++ src/pages/Admin/Core/Templates/index-fund/IndexFundOwner/Form.tsx
@@ -1,6 +1,6 @@
 import { FormValues as FV } from "./types";
-import { FormContainer, Submitter } from "components/admin";
 import { Field } from "components/form";
+import { FormContainer, Submitter } from "../../../../components";
 import useUpdateOwner from "./useUpdateOwner";
 
 export default function Form() {
