+++ src/pages/Admin/Core/Templates/registrar/ConfigExtension/Form.tsx
@@ -1,6 +1,10 @@
 import { FormValues as FV } from "./types";
-import { FormContainer, GroupContainer, Submitter } from "components/admin";
 import { Field, Label } from "components/form";
+import {
+  FormContainer,
+  GroupContainer,
+  Submitter,
+} from "../../../../components";
 import useConfigureRegistrar from "./useConfigureRegistrar";
 
 export default function Form() {
