+++ src/pages/Admin/Core/Templates/registrar/ConfigExtension/index.tsx
@@ -3,7 +3,7 @@ import { FormProvider, useForm } from "react-hook-form";
 import { FormValues } from "./types";
 import { RegistrarConfig, RegistrarConfigUpdate } from "types/contracts";
 import { useContractQuery } from "services/juno";
-import { FormError, FormSkeleton } from "components/admin";
+import { FormError, FormSkeleton } from "../../../../components";
 import Form from "./Form";
 import { schema } from "./schema";
