+++ src/pages/Admin/Proposal/index.tsx
@@ -3,10 +3,10 @@ import { ProposalParams } from "pages/Admin/types";
 import { useProposalDetailsQuery } from "services/juno/custom";
 import QueryLoader from "components/QueryLoader";
 import Seo from "components/Seo";
-import { DetailLabel, Status } from "components/admin";
 import { APP_NAME, DAPP_URL } from "constants/env";
 import { adminRoutes } from "constants/routes";
 import { useAdminContext } from "../Context";
+import { DetailLabel, Status } from "../components";
 import Content from "./Content";
 import PollAction from "./PollAction";
 import Stats from "./Stats";
