+++ src/pages/Admin/Proposals/ProposalCard.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import { Transaction } from "types/contracts/multisig";
-import { Status } from "components/admin";
 import { adminRoutes } from "constants/routes";
+import { Status } from "../components";
 
 export default function ProposalCard(props: Transaction) {
   return (
