+++ src/pages/Admin/components/DetailLabel.tsx
