+++ src/pages/Admin/components/index.ts
