+++ src/pages/Admin/templates/cw3/FundSender/index.tsx
@@ -4,9 +4,9 @@ import { FormValues } from "./types";
 import { Chain } from "types/tx";
 import { useChainQuery } from "services/apes";
 import QueryLoader from "components/QueryLoader";
-import { FormError, FormSkeleton } from "components/admin";
 import { chainIds } from "constants/chainIds";
 import { useAdminContext } from "../../../Context";
+import { FormError, FormSkeleton } from "../../../components";
 import Form from "./Form";
 import { schema } from "./schema";
